remove template slug from body class
The current version of Soil inserts a body class by calling get_page_template().

get_page_template() actually returns page.php when it doesn't find a page template. It does this for everything, including posts. Soil then strips off the .php extension and inserts the class "page" where it doesn't belong. This is particularly problematic if you use .post and .page as CSS selectors, which many people do. The only page that should have the class "page" in WP is a page.

We could add a conditional check to watch for pages, or could use get_template_slug(), which doesn't fall back to page.php here, and only returns a template slug if one is set. However, I propose that we delete it completely. 

WP already outputs the page template slug as a body class, if a page template is set. It is a bit longer, but still can be used, for example page-template-template-name. This function is actually introducing unnecessary duplication in the body class, by also inserting "template-name", without the "page-template" bit. Seems unnecessary to include two classes for the same page template, unless I'm missing the use case here.