Use class full of statics to manage jQuery CDN loading
Bit more code, but I think it's more readable and easier to comprehend. And although it's still tied to jQuery, it would be much easier to begin adding more script fallbacks and filtering more scripts this way.

Although to start filtering more scripts we would probably need to make this an actual class and not all statics.